refactor: move UserData class to types module and update imports across the codebase
This change makes circular imports stemming from the import of this dataclass more unlikely.
