+++ src/mavedb/lib/types/authentication.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from mavedb.models.enums.user_role import UserRole
+    from mavedb.models.user import User
+
+
+@dataclass
+class UserData:
+    user: "User"
+    active_roles: list["UserRole"]
